backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.12

# Optional integrations for full hardware / AI mode:
# anthropic==0.40.0
# faster-whisper==1.0.3
pyserial==3.5
# liblouis bindings vary by platform, so we use a built-in fallback when unavailable.
